Extract policy runner from Trajectory.
